+++ web-client/src/presenter/actions/setModalErrorAction.test.js
@@ -1,0 +1,14 @@
+import { runAction } from 'cerebral/test';
+import { setModalErrorAction } from './setModalErrorAction';
+
+describe('setModalErrorAction', () => {
+  it('should set the modal error message', async () => {
+    const result = await runAction(setModalErrorAction, {
+      props: {
+        error: 'Error message',
+      },
+    });
+
+    expect(result.state.modal.state).toEqual('Error message');
+  });
+});
